requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0
